+++ .gitignore
@@ -18,3 +18,6 @@
 ### Build ###
 /build/
 /dist/
+
+### VSCode ###
+/.vscode
